moved drawing to controller
