+++ main.py
@@ -1,7 +1,6 @@
-import utils.controller.Game as Game
+import utils.controller.game as Game
 import pydantic
 
 game = Game.Game()
-# game.gb.stalemate_test()
 game.main()
